Fix conditional execution for 6.2.6 (root paths)
The "fix paths" task depends on the execution of the previous one, which
collects the paths defined in the sudoers file.

When the first one is skipped by setting the control variable
`ubuntu1804cis_rule_6_2_6` to `false`, the second one failed because the
variable `fixsudo` was not being registered, thus making the play fail
because the `stdout_lines` attribute did not exist.

Embracing what has been done for other sections, we also create a block
for all the related tasks, moving the conditional execution to the block
instead of specifying it for all the tasks.