Misc fixes (#61)
* Fix conditional execution for 6.2.6 (root paths)

The "fix paths" task depends on the execution of the previous one, which
collects the paths defined in the sudoers file.

When the first one is skipped by setting the control variable
`ubuntu1804cis_rule_6_2_6` to `false`, the second one failed because the
variable `fixsudo` was not being registered, thus making the play fail
because the `stdout_lines` attribute did not exist.

Embracing what has been done for other sections, we also create a block
for all the related tasks, moving the conditional execution to the block
instead of specifying it for all the tasks.

* APT package cleanup: fix check mode

If running in check mode, the shell task that gathers the packages in
`rc` state is not executed. As in the cleanup task we are accessing the
registered variable (specifically its `stdout_lines` field), it will
fail.

Change the task to be executed only outside of check mode.

* Section 2.2.10: Fix httpd service name in Ubuntu

In Debian-based distros, the package and service is named `apache2`
instead of `httpd` (RedHat-based distros).

* Section 2.3.5: Fix Ubuntu package name for LDAP clients

* Section 2.2.12: Fix Ubuntu service name for Samba

* Section 2.1.8: Fix Ubuntu service name for Telnet server

* Section 2.1.9: Fix service name for tftp

Also, use systemd as the service manager instead of xinetd.

* Section 2.2.2: Fix package specification for X Window System

In Debian-based distros there is no package group as in RedHat, so here
we change the group for a generic meta-package `xorg`, which is the one
that pulls all the dependencies to install the X Window System.

* Section 2.3.1: Fix package name for ypbind

* Section 2.2.17: Fix service name for ypserv

In Ubuntu, the service name is `nis` instead.

* Section 5.3.1: Fix password quality settings

If running in check mode, the configuration file
`/etc/security/pwquality.conf` does not exist, and by default the
`lineinfile` module fails.

Adding the `create` option allows us to execute the task even when the
file has been removed (or in check mode).

* Section 5.3.3: Password history configurable

Make the maximum password history reusability a configurable value
instead of a hardcoded number. The default reflects the old value.

* Section 4.1.1.1: Fix auditd config in check mode

In check mode, if the configuration file `/etc/audit/auditd.conf` does
not exist, the lineinfile module fails.

This can happen to a first-time apply, when auditd is not installed yet,
but in such cases we don't want to just fail.

* Section 1.8: Allow to leave patching for external tools

A new variable `ubuntu1804cis_apply_upgrades` (default: true) is added
to control whether this role should try to apply upgrades.

Another patching systems may be used, so it may be useful to
leave patching out of this role.